+++ bundle-workflow/src/manifests/dataclass_typechecked.py
