+++ bundle-workflow/tests/tests_manifests/data/invalid-ref.yml
@@ -10,3 +10,4 @@ components:
     checks:
       - gradle:publish
       - gradle:properties:version
+
